Further small fixes

+++ vermeerkat/__init__.py
@@ -8,6 +8,7 @@ import os
 import argparse
 import sys
 import shutil
+from distutils.dir_util import copy_tree
 
 import pkg_resources
 try:
@@ -55,11 +56,6 @@ def create_logger():
     return log, filehandler, console, cfmt
 
 def prompt(prompt_string="Is this configuration correct?", dont_prompt=False):
-    try:
-        input = raw_input
-    except NameError:
-        pass
-
     while not dont_prompt:
         r = input(f"{prompt_string} (Y/N) >> ").lower()
         if r == "y":
@@ -70,21 +66,28 @@ def prompt(prompt_string="Is this configuration correct?", dont_prompt=False):
             continue
     return True
 
-def init_inputdir(INPUT, dont_prompt=False):
-    def __merge_input(INPUT):
+def init_inputdir(INPUT, dont_prompt=False, dont_clean=False):
+    def __merge_input(INPUT, dont_clean=False):
+        if dont_clean:
+            log.warning(f"Attempting to merge pipeline input files into existing '{INPUT}' directory")
+        else:
+            log.info(f"Initializing input directory '{INPUT}' for pipelined run")
         mod_path = os.path.dirname(__file__)
         data_dir = os.path.join(mod_path, "data", "input")
-        shutil.copytree(data_dir, INPUT)
+        copy_tree(data_dir, INPUT)
 
     if not os.path.exists(INPUT):
         __merge_input(INPUT)
     elif os.path.isdir(INPUT):
-        def __reinit(INPUT):
-            shutil.rmtree(INPUT)
-            __merge_input(INPUT)
-        if prompt(prompt_string=f"Input directory '{INPUT}' already exists. Are you sure you want to DELETE and reinitialze?",
+        def __reinit(INPUT, dont_clean=dont_clean):
+            if not dont_clean:
+                shutil.rmtree(INPUT)
+            __merge_input(INPUT, dont_clean=dont_clean)
+        if dont_clean:
+            __reinit(INPUT, dont_clean=dont_clean)
+        elif prompt(prompt_string=f"Input directory '{INPUT}' already exists. Are you sure you want to DELETE and reinitialze?",
                     dont_prompt=dont_prompt):
-            __reinit(INPUT)
+            __reinit(INPUT, dont_clean=dont_clean)
         else:
             log.info("Aborted because the input directory could be initialized per user request")
             sys.exit(1)
